Split hash and project submodule
