+++ gistit-cli/src/fetch.rs
@@ -9,8 +9,9 @@ use url::Url;
 use lib_gistit::file::File;
 use lib_gistit::ipc::{self, Instruction};
 
-use crate::dispatch::{get_runtime_dir, Dispatch, GistitPayload};
-use crate::params::Check;
+use crate::dispatch::{Dispatch, GistitPayload};
+use crate::param::Check;
+use crate::project::runtime_dir;
 use crate::{prettyln, ErrorKind, Result};
 
 lazy_static! {
@@ -83,7 +84,7 @@ impl Dispatch for Action {
     }
 
     async fn dispatch(&'static self, config: Self::InnerData) -> Result<()> {
-        let runtime_dir = get_runtime_dir()?;
+        let runtime_dir = runtime_dir()?;
         let mut bridge = ipc::client(&runtime_dir)?;
         let hash = config.hash;
