+++ gistit-cli/src/hash.rs
@@ -1,0 +1,3 @@
+pub trait Hasheable {
+    fn hash(&self) -> String;
+}
